Refactoring of options widget
The options widget now includes a label giving the stretch value that
auto-updates, and the ability to set that stretch value manually. In
addition, the component of the data can now be selected.

The canvas-specific code setting e.g. camera and data parameters has been
refactored and is now in the vispy widget rather than in the options widget.
The options widget is kept as simple as possible and simply triggers a
refresh from the data viewer when any options are updated.