Improve volume rendering shader.
This is now based on the maximum intensity profile method to avoid having the rendering depend on the depth of the data. This also blends the color using the average (weighted by alpha) and then uses the maximum alpha as the final alpha to prevent pixels from getting *darker* when adding layers (which is counter-intuitive).