change the icon figure
Make sure the 3D viewer is updated if the zorder is set manually. This happens for example when adding a subset to a viewer manually.

Added changelog entry

Raise an explicit warning in 3D volume viewer if PyOpenGL is not present

remove conflicts marks in changes log

fix the change log